skip string read step
