+++ src/setup_runs/cmaq/read_config_cmaq.py
@@ -153,23 +153,7 @@ class CMAQConfig:
     # TODO: Add description for CAMSToCmaqBiasCorrect?
 
 
-def load_json(filepath: str) -> dict[str, str | int | float]:
-    """
-    Loads and parses JSON data from a file.
 
-    Parameters
-    ----------
-    filepath
-        The path to the JSON file to load.
-
-    Returns
-    -------
-        The parsed JSON data.
-    """
-
-    with open(filepath) as f:
-        config = json.load(f)
-    return config
 
 
 def create_cmaq_config_object(config: dict[str, str | int | float]) -> CMAQConfig:
