+++ src/setup_runs/config_read_functions.py
@@ -182,3 +182,21 @@ def process_date_string(datestring: str) -> datetime.datetime:
     date = tz.localize(date)
 
     return date
+
+def load_json(filepath: str) -> dict[str, str | int | float]:
+    """
+    Loads and parses JSON data from a file.
+
+    Parameters
+    ----------
+    filepath
+        The path to the JSON file to load.
+
+    Returns
+    -------
+        The parsed JSON data.
+    """
+
+    with open(filepath) as f:
+        config = json.load(f)
+    return config
